Adds support for specifying objectEncoding for secrets (#236)
Fixes #89

Some secrets are stored encoded in KeyVault (for example, retrieving a pfx certificate from the secrets endpoint returns the base64 encoded certificate).  Many applications will be looking for or expecting the unencoded binary file.

This change adds support for specifying the objectEncoding of a secret.  The supported options are 'hex', 'base64' and 'utf-8'.  The default if not specified is 'utf-8' as that is the current behavior (value from KeyVault is written straight to a file).  If either 'hex' or 'base64' are specified, the secret string will first be decoded before being written to the file.