[Feature] Allow user to specify providersDir
This is needed for Red Hat Container Linux and other immutable hosts
that do not appreciate files being added into `etc`.  It can cause
OpenShift (ARO) to fail to upgrade.

`/var/run` is a more appropriate place to place Linux Sockets, but
changing to that would be a breaking change, this at least allows the
chart deployer to change the path from the default `/etc/kubernetes...`

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>