feat: add functionality for syncing version history (#834)
* feat: add functionality for syncing version history

If specified with a value greater than 1, the ObjectVersionHistory tells
the provider how many versions of the secret (including the latest
version)to sync from key vault. It will then put each secret into a file
named {objectName\objectAlias}/{secretVersion}. To use this
functionality, the provider will need the secrets/list permission in
addition to the permissions it already requires. If this functionality
is not used, then there are no additional permissions needed.